services/training/requirements.txt
# =============================================================================
# MindFu Training Service Dependencies
# =============================================================================

# Note: torch and unsloth are installed separately in Dockerfile

# Transformers ecosystem
transformers>=4.45.0
datasets>=3.0.0
accelerate>=0.34.0
peft>=0.13.0

# Training utilities
trl>=0.11.0
bitsandbytes>=0.44.0
sentencepiece>=0.2.0

# Experiment tracking
mlflow>=2.14.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.35

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.6

# Utilities
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.1
httpx>=0.27.0
tqdm>=4.66.0

# Export
safetensors>=0.4.5
huggingface_hub>=0.25.0
